+++ electronics_lib/Microcontroller_Esp32.py
@@ -249,7 +249,7 @@ class Esp32_Wroom_32(Microcontroller, Radiofrequency, HasEspProgramming, Esp32_I
       self.en_pull = imp.Block(PullupDelayRc(10 * kOhm(tol=0.05), 10*mSecond(tol=0.2))).connected(io=self.ic.chip_pu)
 
 
-class Freenove_Esp32_Wrover(IoControllerUsbOut, IoControllerPowerOut, IoController, Esp32_Ios, GeneratorBlock,
+class Freenove_Esp32_Wrover(IoControllerUsbOut, IoControllerPowerOut, Esp32_Ios, IoController, GeneratorBlock,
                             FootprintBlock):
   """ESP32-WROVER-DEV breakout with camera.
